pyproject.toml
[project]
name = "discordbot"
version = "0.1.0"
description = "A Discord bot built with discord.py"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord.py",
    "python-dotenv",
    "fastapi[standard]>=0.116.1",
]
